fix(db): scope the facility write flag to each RPC call
The transaction-local bypass flag stayed set for the rest of the
transaction after a facility RPC ran, so any later statement in the
same transaction (pgTAP runs a whole file in one) slipped past the
protection triggers. Each RPC now clears the flag before returning,
and the expense-sum assertion casts to bigint for results_eq.

+++ supabase/tests/0016_credit_facilities_installments_test.sql
@@ -230,7 +230,8 @@ select app_finance.create_installment_plan(
 );
 select results_eq(
   $$select
-      (select sum(amount_minor) from app_finance.financial_transactions
+      (select sum(amount_minor)::bigint
+        from app_finance.financial_transactions
         where user_id = '00000000-0000-0000-0000-000000000044'
           and transaction_kind = 'expense'),
       (select total_payable_minor from app_finance.installment_plans
